Add initial layout for base daughterboard
Started the initial layout of the daughterboard with the
microcontroller. The inductors and crystals have placeholder footprints,
while the antenna and ferrite bead are TBD. Currently using default SMD
0402 size, but compact layout is still difficult to achieve. Crystals
have been placed as close as possible to the oscillator pins, but ground
return path is not optimized. Extent of components to be placed on
daughterboard still not finalized.