Add new tests for patch antenna
Two new test boards laid out to test impedance and performance due to
the effect of larger copper void space in the presence of nearby ground
copper, one board with a matching circuit and the other without. Traces
no longer follow the optimal transmission line width for 50ohm
performance due to the use of 0402 SMD technology.