Add USB micro B receptacle footprint
USB B footprint generated for potential data/debugging connection.
XBee could potentially move to bottom right corner of layout by using
UART pins near the bottom of the uC (12-20) and moving the display pins
to the top (36-42). Speaker can be placed directly to the above (to the
right in layout) the display.